fixup node
