+++ bin/node/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "hera"
+name = "node"
 version = "0.1.0"
-description = "Hera is a Rust implementation of the OP Stack Rollup Node"
+description = "A Rust implementation of the OP Stack Rollup Node"
 
 edition.workspace = true
 authors.workspace = true
@@ -13,8 +13,8 @@ rust-version.workspace = true
 
 [dependencies]
 # Local
-hilo-net.workspace = true
-hilo = { workspace = true, features = ["registry"] }
+# hilo-net.workspace = true
+# hilo = { workspace = true, features = ["registry"] }
 
 # Workspace
 eyre.workspace = true
